+++ srcs/parser/parser.c
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/01 20:28:54 by tsishika          #+#    #+#             */
-/*   Updated: 2024/04/02 08:55:23 by tsishika         ###   ########.fr       */
+/*   Updated: 2024/04/03 03:34:27 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ static void	is_set_init(t_is_set *is_set)
 	is_set->light = false;
 }
 
-static check_is_set(t_is_set is_set)
+static void	check_is_set(t_is_set is_set)
 {
 	if (is_set.ambient_lightning == false
 		|| is_set.camera == false
@@ -34,5 +34,5 @@ void	parser(t_data *data, int argc, char **argv)
 	check_argument(argc, argv);
 	check_and_exit_if_directory(argv[1]);
 	input_file(data, argv[1]);
-	check_is_set(is_set);
+	check_is_set(data->is_set);
 }
